pyproject.toml
[project]
name = "experiment-in-computer-vision"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dvc>=3.64.2",
    "mlflow[auth]>=3.7.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "zenml[server]>=0.92.0",
]

[project.optional-dependencies]
vin-ocr = [
    "numpy>=2.1.0",
    "opencv-python>=4.10.0.84",
    "pandas>=2.2.2",
    "pillow>=10.4.0",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.5.2",
    "tqdm>=4.66.5",
]
paddle-ocr = [
    "paddleocr>=3.0.0",
    "paddlepaddle>=3.0.0",
]
hf-ocr = [
    "accelerate>=0.34.2",
    "datasets>=3.0.1",
    "peft>=0.12.0",
    "transformers>=4.44.2",
]
